Bigdata/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
branca==0.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
fastapi==0.110.0
folium==0.15.1
fonttools==4.49.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mysql==0.0.3
mysqlclient==2.2.4
numpy==1.22.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
starlette==0.36.3
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
xyzservices==2023.10.1
zipp==3.17.0
